pyproject.toml
[tool.poetry]
name = "hw-django-skypro"
version = "0.1.0"
description = ""
authors = ["Grokhovskiy Vadim"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
django-bootstrap5 = "^23.3"
pillow = "^10.1.0"
psycopg2 = "^2.9.9"
django-crispy-forms = "^2.1"
crispy-bootstrap4 = "^2023.1"
django-allauth = "^0.60.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
